Add a non-android fallback for vrexternal
