+++ rust-webvr/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust-webvr"
-version = "0.9.13"
+version = "0.9.14"
 authors = ["The Servo Project Developers"]
 
 homepage = "https://github.com/servo/rust-webvr"
